+++ apps/library/services/spreadsheets.py
@@ -5,7 +5,6 @@ from typing import Optional
 from django.conf import settings
 from google.oauth2 import service_account
 from googleapiclient.discovery import build
-from googleapiclient.errors import Error
 
 from apps.main.models import SettingPlaySupply
 
@@ -225,6 +224,6 @@ class GoogleSpreadsheets:
                 self._set_borders(service)
                 self._set_header(service)
             return self._export_new_object(instance, service, file_url)
-        except (ValueError, Error) as error:
+        except (ValueError, Exception) as error:
             msg = f"Не удалось выгрузить данные заявки от {instance.email} на Google Sheets."
             logger.critical(msg, error, exc_info=True)
