+++ config/settings/base.py
@@ -264,3 +264,6 @@ SPECTACULAR_SETTINGS = {
 }
 
 CSRF_FAILURE_VIEW = 'apps.core.views.csrf_failure'
+
+# https://anymail.readthedocs.io/en/stable/installation/?highlight=SERVER_EMAIL#configuring-django-s-email-backend
+SERVER_EMAIL = env("SERVER_EMAIL", default=None)
